backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-socketio==5.10.0
aiohttp==3.9.1
pandas==2.1.4
numpy==1.26.3
email-validator==2.3.0
greenlet>=3.0.0
python-dotenv==1.0.0